Fixes Instance() off-by-one error.
Servers respond with instances separated by double semicolons and
instance elements separated with single semicolons. Some split-foo
was dropping the last semicolon in each instance causing the last
instance attribute to be excluded.

Ex: "ServerName;ILSUNG1;InstanceName;YUKON;tcp;1433" caused tcp's
value of 1433 to not be picked up as an attribute of this response.